+++ .iex.exs
@@ -1,0 +1,29 @@
+import RDF.Sigils
+
+alias RDF.NS
+alias RDF.NS.{RDFS, OWL, SKOS}
+
+alias RDF.{
+  Term,
+  IRI,
+  Literal,
+  BlankNode,
+
+  Triple,
+  Quad,
+  Statement,
+
+  Description,
+  Graph,
+  Dataset,
+
+  XSD
+}
+
+alias RDF.BlankNode, as: BNode
+
+alias RDF.{NTriples, NQuads, Turtle}
+
+alias Decimal, as: D
+
+alias JSON.LD