+++ test/support/test_suite.ex
@@ -7,7 +7,7 @@ defmodule JSON.LD.TestSuite do
 
   def parse_json_file!(file) do
     case File.read(file(file)) do
-      {:ok,   content} -> Poison.Parser.parse!(content)
+      {:ok,   content} -> Jason.decode!(content)
       {:error, reason} -> raise File.Error, path: file, action: "read", reason: reason
     end
   end
@@ -49,7 +49,6 @@ defmodule JSON.LD.TestSuite do
         {:expand_context, file} -> {:expand_context, j(file)}
         option -> option
        end)
-    |> JSON.LD.Options.new
   end
 
   def exception(error) do